Move all modal code into one module
